incorporated review
