+++ packages/core/src/reducers/validation.ts
@@ -62,3 +62,8 @@ export const validationReducer = (
 export const errorAt = instancePath => (state): any[] => {
   return _.filter(state.errors, (error: ErrorObject) => error.dataPath === instancePath);
 };
+export const subErrorsAt = instancePath => (state): any[] => {
+  const path = instancePath + '.';
+
+  return _.filter(state.errors, (error: ErrorObject) => error.dataPath.startsWith(path));
+};
