+++ packages/material/src/complex/index.ts
@@ -1,1 +1,11 @@
-export * from './table-array.control';
+import {
+    isArrayObjectControl,
+    rankWith,
+    registerStartupRenderer,
+} from '@jsonforms/core';
+import MaterialArrayControlRenderer from './MaterialArrayControlRenderer';
+
+export default registerStartupRenderer(
+    rankWith(3, isArrayObjectControl),
+    MaterialArrayControlRenderer
+);
