+++ packages/vanilla/src/additional/array/index.ts
@@ -1,17 +1,17 @@
 import {
+  isArrayObjectControl,
+  rankWith,
   registerStartupRenderer,
 } from '@jsonforms/core';
 import ArrayControlRenderer from './ArrayControlRenderer';
 import { ArrayControl } from './ArrayControl';
-import { arrayTester } from './tester';
 
 export {
-  arrayTester,
   ArrayControl,
   ArrayControlRenderer
 };
 
 export default registerStartupRenderer(
-  arrayTester,
+  rankWith(2, isArrayObjectControl),
   ArrayControlRenderer
 );
